+++ src/main/java/com/popcorntalk/domain/point/repository/PointRepository.java
@@ -4,7 +4,7 @@ import com.popcorntalk.domain.point.entity.Point;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointRepository extends JpaRepository<Point, Integer> {
+public interface PointRepository extends JpaRepository<Point, Long> {
 
     Optional<Point> findByUserId(Long id);
